% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_convsummary.R
\name{get_convsummary}
\alias{get_convsummary}
\title{Convergence issue summary}
\usage{
get_convsummary(object)
}
\arguments{
\item{object}{A dynforest object.}
}
\value{
A data.frame summarising convergence issues by variable.
}
\description{
Convergence issue summary
}
